requirements.txt
celery==5.2.3
fastapi==0.73.0
ipdata==3.4.3
peewee==3.14.8
pydantic==1.9.0
python-dotenv~=0.19.2
PyMySQL==1.0.2
